[Entity Store] Move asset manager to asset manager client
