+++ x-pack/solutions/security/plugins/entity_store/server/domain/asset_manager/component_templates.ts
