+++ x-pack/solutions/security/plugins/entity_store/server/domain/asset_manager/external_indices_contants.ts
