+++ x-pack/solutions/security/plugins/entity_store/server/domain/asset_manager/history_snapshot_index.ts
