+++ x-pack/solutions/security/plugins/entity_store/server/domain/asset_manager/latest_index_template.ts
