+++ x-pack/solutions/security/plugins/entity_store/server/domain/asset_manager/updates_index_template.ts
