+++ x-pack/solutions/security/plugins/entity_store/server/routes/apis/uninstall.ts
@@ -37,7 +37,7 @@ export function registerUninstall(router: EntityStorePluginRouter) {
         },
       },
       wrapMiddlewares(async (ctx, req, res) => {
-        const { logger, assetManager } = await ctx.entityStore;
+        const { logger, assetManagerClient: assetManager } = await ctx.entityStore;
         const { entityTypes } = req.body;
         logger.debug(`uninstalling entities: [${entityTypes.join(', ')}]`);
