+++ x-pack/solutions/security/plugins/entity_store/server/types.ts
@@ -24,7 +24,7 @@ import type { Logger } from '@kbn/logging';
 import type { SpacesPluginSetup, SpacesPluginStart } from '@kbn/spaces-plugin/server';
 import type { CoreSetup } from '@kbn/core/server';
 import type { ElasticsearchClient } from '@kbn/core/server';
-import type { AssetManager } from './domain/asset_manager';
+import type { AssetManagerClient } from './domain/asset_manager/asset_manager_client';
 import type { EntityMaintainersClient } from './domain/entity_maintainers';
 import type { FeatureFlags } from './infra/feature_flags';
 import type { CcsLogsExtractionClient, LogsExtractionClient } from './domain/logs_extraction';
@@ -49,7 +49,7 @@ export interface EntityStoreStartPlugins {
 export interface EntityStoreApiRequestHandlerContext {
   core: CoreRequestHandlerContext;
   logger: Logger;
-  assetManager: AssetManager;
+  assetManagerClient: AssetManagerClient;
   entityMaintainersClient: EntityMaintainersClient;
   crudClient: CRUDClient;
   ccsLogsExtractionClient: CcsLogsExtractionClient;
